chore: moved upgrades into src folder, imported header files into main.cpp
