+++ src/upgrades/hideBot.h
@@ -1,7 +1,7 @@
 #ifndef HIDEBOT_H
 #define HIDEBOT_H
 
-#include "../src/genericRobot.h"
+#include "../genericRobot.h"
 
 class HideBot:public GenericRobot{
     public:
