+++ src/upgrades/jumpBot.h
@@ -1,7 +1,7 @@
 #ifndef JUMPBOT_H
 #define JUMPBOT_H
 
-#include "../src/genericRobot.h"
+#include "../genericRobot.h"
 class JumpBot:public GenericRobot{
     public:
         void jump(int x, int y);
