+++ src/upgrades/semiAutoBot.h
@@ -1,9 +1,9 @@
 #ifndef SEMIAUTOBOT_H
 #define SEMIAUTOBOT_H
 
-#include "../src/genericRobot.h"
+#include "../genericRobot.h"
 
-class ThirtyShotBot:public GenericRobot{
+class SemiAutoBot:public GenericRobot{
     public:
         void fire(int x, int y) override;
 };
